requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tiktoken>=0.5.0
sqlite3
datetime
json
uuid
pathlib
asyncio
typing
logging
